insaerospace/requirements.txt
requests
python-dotenv
django
gunicorn